+++ zappend/context.py
@@ -2,6 +2,7 @@
 # Permissions are hereby granted under the terms of the MIT License:
 # https://opensource.org/licenses/MIT.
 
+import inspect
 import tempfile
 from typing import Any, Dict
 
@@ -48,6 +49,11 @@ class Context:
         temp_dir_uri = config.get("temp_dir", tempfile.gettempdir())
         temp_storage_options = config.get("temp_storage_options")
         self._temp_dir = FileObj(temp_dir_uri, storage_options=temp_storage_options)
+
+        from .slice.factory import to_slice_source_type
+
+        slice_source = config.get("slice_source")
+        self._slice_source = to_slice_source_type(slice_source)
 
     def get_dataset_metadata(self, dataset: xr.Dataset) -> DatasetMetadata:
         """Get the dataset metadata from configuration and the given dataset.
@@ -98,8 +104,15 @@ class Context:
         return self._config.get("slice_engine")
 
     @property
+    def slice_source(self) -> Any | None:
+        """The configured slice source, if any."""
+        return self._slice_source
+
+    @property
     def slice_storage_options(self) -> dict[str, Any] | None:
-        """The configured slice storage options to be used if a slice object is a Zarr."""
+        """The configured slice storage options to be used
+        if a slice object is a Zarr.
+        """
         return self._config.get("slice_storage_options")
 
     @property