+++ zappend/processor.py
@@ -16,7 +16,7 @@ from .log import logger
 from .log import configure_logging
 from .rollbackstore import RollbackStore
 from .slice import SliceObj
-from .slice import get_slice_dataset
+from .slice import open_slice_source
 from .tailoring import tailor_target_dataset
 from .tailoring import tailor_slice_dataset
 
@@ -83,7 +83,7 @@ class Processor:
 
         ctx = Context(self._config)
 
-        with get_slice_dataset(ctx, slice_obj, slice_index) as slice_dataset:
+        with open_slice_source(ctx, slice_obj, slice_index) as slice_dataset:
             slice_metadata = ctx.get_dataset_metadata(slice_dataset)
             if ctx.target_metadata is None:
                 ctx.target_metadata = slice_metadata